eric-am-shared-spring-security/src/main/java/com/ericsson/am/shared/spring/security/utils/AuthenticationUtils.java
/*
 * COPYRIGHT Ericsson 2024
 *
 *
 *
 * The copyright to the computer program(s) herein is the property of
 *
 * Ericsson Inc. The programs may be used and/or copied only with written
 *
 * permission from Ericsson Inc. or in accordance with the terms and
 *
 * conditions stipulated in the agreement/contract under which the
 *
 * program(s) have been supplied.
 */
package com.ericsson.am.shared.spring.security.utils;

import java.util.Optional;

import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
import org.springframework.security.core.Authentication;
import org.springframework.security.core.context.SecurityContextHolder;

public final class AuthenticationUtils {
    private AuthenticationUtils() {
        throw new UnsupportedOperationException("Utility class should not be instantiated");
    }

    public static Optional<String> getUserName() {
        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
        return Optional.ofNullable(authentication)
                .filter(auth -> auth instanceof UsernamePasswordAuthenticationToken)
                .map(Authentication::getName);
    }
}
